[buildbucket][bq-export] improve non-finalized build handling
bq.py reads chains of properties in a nested dicts. Some of them may be
present and be None. Handle that.

Treat TIMED_OUT same way as BOT_DIED: in both cases, it is OK if steps
are not finalized.

Reduce waiting threshold from 1h to 20m.

R=tandrii@chromium.org

Bug: 742478
Change-Id: I5479f7f6d7cb1e2ab2468cbba19fcdff5e4eb376
Reviewed-on: https://chromium-review.googlesource.com/952780
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Ryan Tseng <hinoka@chromium.org>