Send a run details page URL with robot comments
In robot comments in Gerrit, there's already a "robot run link",
but it doesn't link to anything because no URL is being added to
comments. This would associate a URL with a run and allow Gerrit
to make links to details for Tricium runs.

Bug: 820118
Change-Id: I7073829a593c301c09164b7e7c234970d0d2e597
Reviewed-on: https://chromium-review.googlesource.com/956527
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>