[recipe_bundler] Make prefixes required, always set-ref.
This makes the prefix settings explicitly required by the tool to prevent
accidental uploads to the production CIPD packages.

It also changes so that when the tool is supposed to check out "FETCH_HEAD"
(aka "the latest version of the given ref"), it will call `cipd set-ref`
even if the CIPD package tagged with the resolved git commit has already
been uploaded.

R=vadimsh@google.com

Bug: 811974
Change-Id: I3e0b7b0db3f9200f26d59846afbb2e6e46cd2d34
Reviewed-on: https://chromium-review.googlesource.com/956880
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>