[Predator] Only update CrashAnalysis in datastore, when a new analysis is needed.
Fracas broke because we reinitialize crash analysis even when we don't need a new analysis (the crash is analyzed before). And after we reset all the results, we publish empty results to Fracas.

Fix this error by only updating CrashAnalysis in datastore when we need a new analysis.

The reason why this error didn't impact Clusterfuzz and Cracas is that:
Predator redo the analysis for all the Clusterfuzz and Cracas crashes. so

    need_analysis = predator_client.NeedsNewAnalysis(crash_data)
    if need_analysis:
      logging.info('New %s analysis is scheduled for %s',
                   self._client_id, crash_data.identifiers)

      UpdateCrashAnalysisData(crash_data, predator_client)
      run_analysis = yield CrashAnalysisPipeline(self._client_id,
                                                 crash_data.identifiers)
      with pipeline.After(run_analysis):
        yield PublishResultPipeline(self._client_id, crash_data.identifiers)
    else:
      yield PublishResultPipeline(self._client_id, crash_data.identifiers)

The need_analysis is always true, so we cannot hit the else clause and trigger the error.

Change-Id: Ib19b9815b47639a18383a87420d19eb66f58f8df
Reviewed-on: https://chromium-review.googlesource.com/957866
Commit-Queue: Sharu Jiang <katesonia@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>