[luci-migration] always display Status
Pages don't tell which CI builders are migrated, i.e. deleted from
Buildbot. Restore Status column on master page for CI masters and display status on builder
page. For CI builders, this column can values "" or "Migrated", not others.

Display unknown status in table cells as "".

Also, fix a bug: if a builder was migrated, don't display it as outdated. Also, it can
be outdated only if it is a try builder.

R=estaab@chromium.org

Bug: 820687
Change-Id: Ia3491594a53f7c8701d05867b83a77c1d1c1ba29
Reviewed-on: https://chromium-review.googlesource.com/959235
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Erik Staab <estaab@chromium.org>