For internal routes, set up auth state using anonymous identity
This is the fix proposed by Nodir for the issue that State
wasn't set in the context. The reason was because normally
that State is set up by Middleware, but the MiddlewareChain
for internal routes (including the poll cron job) didn't
include anything that sets up State.

This CL adds a Middleware auth.Authenticate that sets the
current identity as anonymous and doesn't do anything else.
This is OK for the poller since the current identity isn't
checked; it's the CL owner's identity that's checked.

As a follow-up, I'd like to:

 1. Add "anonymous" auth method in the relevant package in luci-go
    (maybe "go.chromium.org/luci/appengine/gaeauth/server"?)
 2. Change this code to use that.

I tested this by deploying a dirty version for a minute in
the evening, and it appeared to work as expected.

Bug: 825521
Change-Id: I7f49fca10545517bbf798d75b593181688fcfbe7
Reviewed-on: https://chromium-review.googlesource.com/985392
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>