[Findit] Decouple cl_key from repo/revision
When a pipeline refers to a suspected cl, it works by calling Get with
the required properties. Since the class structure is now:
Base --> (Derived_1, Derived_2)
There's no way to use these properties to get the proper derived model.
This change introduces a string key that will be passed instead of
the usual (repo, revision). This will allow for use with each of the
derived model classes.

Notable changes:
* Removed the CLKey structured object
* Kept the cl_key property, but it's now a basestring
* Changed references to accommodate.
* Created wrapper method to handle potential errors when getting a
  bad key (this is sort of an open issue for ndb in general).
  https://github.com/googlecloudplatform/datastore-ndb-python/issues/143

Bug: 826870
Change-Id: I9cfaeb635d46eee3302a75c1d27ecd6044c114fe
Reviewed-on: https://chromium-review.googlesource.com/985194
Commit-Queue: Brandon Wylie <wylieb@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>