[cr-audit-commits] Add NotificationTokens field to RuleResult.
This change is meant to allow rules pass arbitrary string tokens to the
notification functions, that are not included in the message.

Note that the reason why we use a single string is because of limits
of the datastore library. There is a restriction about nested structs
(like RuleResult) containing only flat fields.

R=chanli,stgao,cmasso

Change-Id: Ic705023c2da8151b2be65dc825e06e17c2dc3eab
Reviewed-on: https://chromium-review.googlesource.com/982756
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>