[cr-audit-commits] Remove release-branch specific field MilestoneNumber.
This change removes a field from the RepoConfig that is only applicable
to release branches, instead, it is replaced with a more flexible
"Metadata" field that can be used to store arbitrary data specific to
the repository configuration.

This change exemplifies how to use the utility function GetToken to read
this data from the RepoConfig struct.

R=stgao,cmasso

Change-Id: Ie22cbe4ac821ff094480a707e533092c8c88eb25
Reviewed-on: https://chromium-review.googlesource.com/998342
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>