[Monorail] Support user-defined query strings for snapshot counts
This adds a URL query parameter q= to the snapshot counts endpoint. It
accepts the same type of query string as the q= parameter in the issue list
UI, but only supports a subset of the tables and conditions based on what
snapshots we have stored. The names of any unsupported fields are returned
in the response along with the snapshot counts.

Bug: monorail:3534
Change-Id: Ic9e8c75d683b71ebbaff3ccfe208b8ef8d7ef8e2
Reviewed-on: https://chromium-review.googlesource.com/984406
Commit-Queue: Jeff Carpenter <jeffcarp@chromium.org>
Reviewed-by: Jason Robbins <jrobbins@chromium.org>