Allow star to mean "don't filter by CL owner" again
This is an optional change that would make it so that
we support both "*" and absence of whitelisted groups
to mean no filtering.

Committing this is optional, but if we commit this
temporary change then we don't necessarily need to
update all project configs before starting to read
project details from project configs.

I plan to revert this CL after some time once stars
are removed from configs.

Change-Id: I335030d3760f29ce3f386859733e4402da3efe91
Reviewed-on: https://chromium-review.googlesource.com/1005436
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>