[Monorail] Refactor MonorailContext to framework.
My goal here is to separate the request mechanics in the
mc from the request parsing in the mr for both UI and API
code paths so that shared code in work_env can just take a
mc.  In this step, it will be a matter of accepting a base
class vs a subclass, but eventually they will be separate
objects.

R=agable@chromium.org

Bug: monorail:3599
Change-Id: I4701287c40c03030c9eb8f040b0870261ed92773
Reviewed-on: https://chromium-review.googlesource.com/999166
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Aaron Gable <agable@chromium.org>