[Findit] Handle edge cases for bounding builds of a commit position.
If a given commit position is included in a boundary build, return that build
as both the lower and upper bound build.

The use case in regression_range_analysis_pipeline.py is not verified to be supported by this change (although most likely be supported), because that feature is not working in fact.

Change-Id: I678f4b1efd0bf2d2fd830634ef4311fc12f5764e
Reviewed-on: https://chromium-review.googlesource.com/1009072
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Shuotao Gao <stgao@chromium.org>