[cr-audit-commits] Support auditing Findit auto-reverts for flake.
This change bypasses one of the rules because the build with the first
failure of the affected test may not contain the culprit of the
flakiness, as it may be older than those in the build.

This change relies on the prefix "Flaky test name: " used by Findit to
indicate the name of the test that is flaky, being present in the commit
message of such auto-reverts to determine if they are meant to address a
flake and thus bypass this rule.

R=wylieb,chanli,stgao,lijeffrey
BUG=831880

Change-Id: I1a3f3964d5fc5fb302935798b95967b134ba7891
Reviewed-on: https://chromium-review.googlesource.com/1008983
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Reviewed-by: Brandon Wylie <wylieb@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>