[Monorail] Avoid invalidating the project for recent activity time.
We have a feature where the recent activity timestamp is shown for
each project on the project list page.  To update that timestamp, we
were updating the project and invalidating it on each issue creation
and edit.  The problem with that is that it makes all our handlers
more likely to get a cache miss on the project object.  This change
makes the recent timestamp granual one hour, probably reducing
project invalidations by 98%.

R=seanmccullough@chromium.org

Bug: monorail:3678
Change-Id: Ic554125d033060f8191a7dd37dec570ad854fd53
Reviewed-on: https://chromium-review.googlesource.com/1014287
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>