ChOpsUI: Revamp ChOpsUI gallery site.
Change-Id: Ief8c19eb812ceb371ab181cb7b3e20240a72ce44

Demo: https://14969-2971eb3-tainted-zhangtiff-dot-chopsui-gallery.appspot.com/

This CL replaces iron-component-page on chopsui-gallery with raw iron-doc-* elements, which allows us to customize the site a lot more. I'm still not totally happy with the amount of customization I was able to do with these particular elements, so I might go a layer deeper in the future, but I figured I'd save that for the future.

Notable changes:

Tried to change the design of chopsui-gallery into something that felt more on-brand for ChOps.
Added a way to expose Markdown doc pages on the site, so we can start turning the site into a central hub of ChOps frontend docs if desired.
Replaced iron-component-page  with iron-doc-nav and iron-doc-viewer (some slightly hackish styling was needed, unfortunately, to make the elements around these components match)
Added a bit more info to the ChOps style guide that Jeef started. (Thanks, Jeff!)

Change-Id: Ief8c19eb812ceb371ab181cb7b3e20240a72ce44
Reviewed-on: https://chromium-review.googlesource.com/1013388
Commit-Queue: Tiffany Zhang <zhangtiff@chromium.org>
Reviewed-by: Jeff Carpenter <jeffcarp@chromium.org>