Change analyzer lookup to return first implementation by default
Reason: In project configs, we currently have a lot of:

  selections {
    function: ...
    platform: UBUNTU
  }

where we don't actually care about about the platform, since all
the analyzers so far are platform-independent.

If no platform is specified, then the platform field is 0, which
corresponds to Platform_ANY. If we're trying to get an implementation
for ANY platform, we could just return the first.

Bug: 824948
Change-Id: Ib4c040540198d90a5ee6a1bbf20f90d420dbfd0b
Reviewed-on: https://chromium-review.googlesource.com/1013092
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>