[cr-audit-commits] Use a taskqueue to run audits.
Also, use a single cron job to schedule the audit tasks.

To be clear, this new approach involves two phases (Scheduler Cron Job,
and Auditor Task Queue)

Phase I - Scheduler Cron Job
A new handler called Scheduler is added to iterate over the available
configs and schedule needed audit tasks, creating new RepoState
datastore entries as needed.  Note that the URL to each ref is passed
to the taskqueue task (Auditor handler below), and that the name of the
corresponding config has been added as a field to the RepoState entity.

Phase II - Auditor Task Queue
This change replaces the CommitScanner and CommitAuditor handlers with a
single handler called Auditor, intended to be run as a taskqueue task
when scheduled by the 'Scheduler' cron job above. This task will scan
the git ref, whose url is given as a parameter and using the matching
configuration (from the hardcoded RuleMap) it will create RelevantCommit
entities when appropriate, followed by an audit of such commits and
notifications for those that are found to be in violation of any of the
rules.

The following design docs contain the discussions about this change:
   - go/audit-configuration-for-dynamic-refs
   - go/syncing-audit-app-scheduler-with-audit-task

R=stgao,chanli
BUG=803314

Change-Id: I1b7f52fa68b41db9cc5575f2a42041afde36443c
Reviewed-on: https://chromium-review.googlesource.com/1010976
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>