[Findit] Flake Analyzer - Fixing key errors in handling swarming task pipeline timeout
When the pipeline for a swarming task times out waiting for the task to trigger or
complete, not all information is available. The calling code expects task_id,
created_ts, etc. to be populated, but they may simply not be available yet depending
on the timeout leading to key errors. Because of the key errors, the callback task to
raise a pipeline abort does not have the chance to abort the asynchronous pipeline
causing the pipeline to appear as though it's still running after the timeout.

Bug: 833455

Change-Id: I0467575729947b3ee86ac4266b55f4ac55a1818f
Reviewed-on: https://chromium-review.googlesource.com/1020463
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>