Change PathFilter matching to apply only to file basenames
Path filters are matched against paths with filepath.Match;
this matching works like Bash glob matching, so
foo/*.md matches "foo/bar.md", but "*.md" does not.

But, for analyzer path filters, we mainly want to use path
filters to be able to express "this analyzer applies to a
certain type of file.".

This CL changes the behavior so that "*.ext" matches files
with that extension regardless of directory nesting.

Bug: 832920
Change-Id: I3c7456303fcd5ff4bff2bd6e1ba949bd9aa1f53a
Reviewed-on: https://chromium-review.googlesource.com/1023595
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>