Don't include binary files in paths to analyze
This is a simple and possibily temporary change which would make
it so that Tricium doesn't look at binary files (e.g. images) in CLs.

We may want analyzers to be able to look at images and other binary
files, but if we are to include both binary and text files, we want
to provide an easy way for analyzers to tell which files are text.

Bug: 836548
Change-Id: I98b80f0ee1b3bf81dc241884791cebf89350e376
Reviewed-on: https://chromium-review.googlesource.com/1028369
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>