Change AnalyzeRequest to include GitRepo
Background:

We'd like to support multiple repositories per project config,
because this will generally be better for fuchsia (and possibly other
projects); it may make the way to configure things a little more
flexible, so you could have one config (which can include analyzer
definitions and selections) apply to a set of related configs
(I'm also thinking of infra repos here).

As part of https://crbug.com/824558, this change to the config
has now already been made.

But, if there can be more than one repo per project, then the
project name alone is not enough to get the repo details; we also
need to know which repo.  For example, when processing an analyze
request, we need to get the git repo so that we can make the initial
GIT_FILE_DETAILS data type.

All of this only applies to Git, since Git is the only supported
repo type currently.


This change proposes to make it so that AnalyzeRequest also
includes GitRepo, which should match the Repository field in
GitRepoDetails. The Gerrit poller would then use this repo URL
which making analyze requests. So, the poller will require both
git details and gerrit details in order to poll for a repository.

Bug: 824558
Change-Id: I61271d918d7683ebadcf22a526c6b4430df59a86
Reviewed-on: https://chromium-review.googlesource.com/1033796
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>