[Monorail] Stop passing seq_num when queueing notifications.
This CL does parts 2 and 3 of issue monorail:3550.  We now identify
the comment to notify a user about by using the comment ID rather than
the comment sequence number.  This finishes a change to avoid a race
condition that can happen when two comments are added to the same
issue rapidly.

R=jojwang@chromium.org

Bug: monorail:3550
Change-Id: Ia06e15aaddd35bc359cab18b740eac127295f5b1
Reviewed-on: https://chromium-review.googlesource.com/1033801
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Joanna Wang <jojwang@chromium.org>