[Monorail] Detect and work-around replication lag when loading comments.
We don't just retireve comment rows, because that can be very slow for
issues that have a huge number of comments.  Instead we retrieve
"abbriviated" rows that contain the comment ID, then paginate those,
then retrieve only the comments that will be shown on the page.

In this CL:
+ Always retrieve abbriviated comment rows from the master, so that
  there can never be replication lag.
+ When retrieveing the full comments from a replica, check that the
  replica has all the comment rows that we saw on the master, otherwise
  load the comment data from the master.
+ Load approval rows from a shard just like other comment details.
+ Fix a call to get all comment in issuepeek (not currently reachable).

R=jeffcarp@chromium.org

Bug: monorail:3755
Change-Id: I7f7b14a603b648450d14954df1f94f18d3483c02
Reviewed-on: https://chromium-review.googlesource.com/1038653
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Jeff Carpenter <jeffcarp@chromium.org>