[Findit] Flake Analyzer - Removing transactional from updating data points
There is no need for this function to be transactional, which is also causing
trouble. @transactional is to guard against multiple analyses trying to
write to the same ndb entity. In the case of Flake Analyzer, there is only 1
active analysis for a given flaky test at a time, thus @transactional provides
no benefit. In the existing code when populating data with build level
information, step_util relies on wf_config, which asserts the calling code
cannot be in a transaction, causing this function to fail. A follow-up CL not
to retrieve build-level information here will come shortly.

Bug: 838326
Change-Id: I6d6a1d26d3d8cd816a1d4dd5edf1d585ec9379cf
Reviewed-on: https://chromium-review.googlesource.com/1040270
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>