[cr-audit-commits] Use Ref URL as key to status page.
After the changes to support dynamic refs, the status page logic wasn't
updated to support the new logic where a single config could apply to
multiple refs.

This change makes the status page use the url to the ref as the
parameter that chooses which ref to display information for, instead of
the previously used repo name.

R=stgao

Change-Id: Iec6103a98a86e6deee986e07ad35e26d02788cf3
Reviewed-on: https://chromium-review.googlesource.com/1043393
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>