[Monorail] Speed up construction of issue view objects.
This basically changes some unneccessarily, nested for-loops into a
series of unnested for-loops.

In staging, the old code took 36ms to 42ms for this phase.  The new code
reduces that to 30ms to 37ms.  In prod, the old code was often over 100ms.
I'm expecting savings of at least 30ms in prod.  That savings would apply
to every issue detail request because this does not depend on cacheing.

R=seanmccullough@chromium.org

Bug: monorail:3801
Change-Id: I858e7af9da42e75cdcafcafdd0aae0916474c5e5
Reviewed-on: https://chromium-review.googlesource.com/1047852
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>