[buildbucket] Do not remove "<key>:" dimensions
Given dimension lists

    d1 = ["a:1"]
    d2 = ["a:"]
    d3 = merge(d1, d2)

d3's a must be empty. However, with

    d4 = ["a:2"]
    d5 = merge(d4, d3)

d5's a must be empty too. If d3 is empty, it is not the case.

Do not exclude empty values from merged dimensions. Instead, ignore them when
interpreting dimensions.

Also fix parsing of dimensions in properties of the rendered task template.
Also do not sort lists of dicts in tests. This is not very predictable.

Bug: 836626
Change-Id: I874f297f1971b1450fdea9a13bbe3d9c20afe26c
Reviewed-on: https://chromium-review.googlesource.com/1048119
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>