[tricium] Replace RunInParallel with FanOutIn
RunInParallel would take a slice of `func() error`, run them in parallel, and
return an error if any of the functions returned an error. FanOutIn is similar,
but it takes a function that takes a channel, which the `func() error` are sent
into.

The advantage of replacing this is allows us to remove RunInParallel, reducing
the total amount of code by a little bit, and this may also make Tricium more
similar to other code that uses luci-go libraries.

Bug: 838644
Change-Id: Iac44566b0a70e77a16ab22e4d1d6cf5d87426b1c
Reviewed-on: https://chromium-review.googlesource.com/1048766
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>