[Findit] Flake Analyzer - Adding upper bound build number as get isolate sha input
When determining nearby builds to a build number, an optional parameter is supplied
for the latest build number to check. For flake analysis, the analysis' triggering
build number is fine, but when verifying flakiness in a recent build that value will
be out dated. This change introduces a parameter for the calling code to specify the
upper bound instead of relying on the analysis' pre-stored value.

Bug: 841638
Change-Id: I7462faac2db5b81ce2e5f830baa886e7f8e1031a
Reviewed-on: https://chromium-review.googlesource.com/1053422
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>