[Findit] Flake Analyzer - Allow for analyses to be reverted as long as they have bugs
The root cause is the field MasterFlakeAnalysis.has_filed_bug is only True if the
analysis itself was the one responsible for filing the bug, and not that any bug is
associated with the analysis. This should not be checked for in auto revert, as multiple
analyses can reach the same conclusion, and only 1 of them is the one that files the bug.
If that analysis decides not to auto revert, no analysis can auto revert subsequently even
if they come to more drastic conclusions (e.g. first one bails out due to insufficient
confidence, but a second analysis detects the same CL introduced a new flaky test and has
higher confidence). This field also doesn't work if the bug was filed by CTF.

This change switches to see if any bug is associated with the analysis, and not that the
particular analysis was responsible for filing it.

Bug: 841920
Change-Id: Ib7fec9664a2db3e8849a1a3697c6179f6f0d0a1d
Reviewed-on: https://chromium-review.googlesource.com/1054410
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Commit-Queue: Shuotao Gao <stgao@chromium.org>