[tricium] Change config proto field numbers to avoid re-use
Vadim noted in the bug that binary protos for configs
are actually stored/cached, which means that changing
the numbers associated with fields will break things
(at least temporarily).

This changes the config back so that it uses the same
numbers for the same fields as before and new numbers
for new fields. This CL also removes a comment that
should have been removed before.

Bug: 824558
Change-Id: I458c2c2a7257ad196ed32925e15b9da75a769959
Reviewed-on: https://chromium-review.googlesource.com/1056206
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>