[Monorail] Fix SQL whitelisting after table alias change.
We changed the name of the table aliases from "User1" to "Spare1"
because it was not just for the User table.  However, we did not
update the SQL whitelist to allow those SQL statements to be executed.

Unrelated: fixed a mismatched method name to clear up pylint output.

R=jojwang@chromium.org

Bug: monorail:3828
Change-Id: Iff0b0b74fd5e2ab9bdc1486cf33695eca3293e6e
Reviewed-on: https://chromium-review.googlesource.com/1060292
Reviewed-by: Joanna Wang <jojwang@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>