[Monorail] Implement permission checks in work_env.
The main part of this change is the new permission checking functions
and lines in the existing methods of work_env.

The new permission-checking and input validation semantics of work_env
requires that some callers were updated.  E.g., loading the project
in monorailrequest cannot go through work_env because we need to get
the project before we calculate permissions.  E.g., when a regular user
posts an attachment, they cannot use the full UpdateProject() method
because they are not a project owner.  Also, more auth info needed to
be added to inboundemail.py and several unit tests.

Bug: monorail:2985
Change-Id: I14cecf1c366749c5196475ee27fa65650d2b5ce7
Reviewed-on: https://chromium-review.googlesource.com/1045545
Reviewed-by: Joanna Wang <jojwang@chromium.org>
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>