Monorail - fixed a tracker.sql script exception.
Running tracker.sql in certain MySQL editions/versions fails
with an error -
Issue2FieldValue - foreign key constraint is incorrectly formed

show warnings showed that the problem was that IssuePhaseDef,
a table that is referenced by the foreign key constraint of
Issue2FieldValue, could not be found. This happened because
it was not created at that point - it would have been created
much later in the script. The fix is to move the creation of
IssuePhaseDef right before the creation of Issue2FieldValue.

R=jrobbins@chromium.org

Bug: monorail: none
Change-Id: Ic2119d9075d1c43793ef265d7feb3cc904ceb902
Reviewed-on: https://chromium-review.googlesource.com/1064513
Reviewed-by: Jason Robbins <jrobbins@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>