[Findit] Flake Analyzer - Temporary workaround to get latest regression range
The root cause of the assertion firing is the analysis had likely performed
a check for flakiness in a recent build, which came back stable. This data
point should never have been appended to begin with, as it interferes with
the rest of lookback_algorithm and the function to get the latest regression
range, etc.

The change not to append data points is much more involved, so as a temporary
workaround, disregard that data point when calculating the latest regression
range.

Bug: 843848
Change-Id: I88d775f0f8e1964b9fc852ad281d05ba4cd3d342
Reviewed-on: https://chromium-review.googlesource.com/1063255
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>