[Monorail] Kill boilerplate, even if decorator changes signature.
This is a follow-up to my CL
https://chromium-review.googlesource.com/c/infra/infra/+/979216 that
introduced the Run() method and Do*() methods.  Instead of boilerplate
methods that call Do*() methods, we will now use decorators.
Previously, I had not liked the idea of a decorator that changed the
method signature, but as I implement more, the boilerplate is proving
to be too tedious.

R=jojwang@chromium.org

Bug: monorail:3599
Change-Id: Ia80d667435f09390e90d619faed421b3d2002ee0
Reviewed-on: https://chromium-review.googlesource.com/1067002
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Joanna Wang <jojwang@chromium.org>