[WPT import] Create a dummy home directory for importer
`git cl` tries to use the cached cookies (in the home directory) by
default, which is the GCE service account, regardless of the git
user.email setting. This CL creates a dummy home directory without the
cached cookies for the importer so that the `git cl` invocation will
look up the token for the importer account from .netrc.

One benefit of using a dummy HOME instead of removing the cached cookies
is to prevent potential unintended effects on other processes or runs.

Bug: 844563
Change-Id: Idf8426b478686601ccdf41ca21b24dca577d4101
Reviewed-on: https://chromium-review.googlesource.com/1070591
Commit-Queue: Robert Ma <robertma@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>