[Monorail] Faster filtering of service accts from autocomplete.
We were looping over each member, and for each constructing two lists
for the API whitelist, then checking membership in one of those lists.
Now we get the whitelist once, convert it to a set, and check membership
much faster.

On staging, this seems to speed up issueOptionsMembers by about 1s
(which is about half of the time that is spent on anything other than
retrieving User objects).

R=seanmccullough@chromium.org

Bug: monorail:3724
Change-Id: I405cd20836bcae6e33428125cab9bd85104bf5e2
Reviewed-on: https://chromium-review.googlesource.com/1072495
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>