[cipd] Make the backend understand ProcessingResult written by Go version.
Golang GAE SDK has no notion of compressed fields. We can still compress the
field manually and put it into the entity, but it doesn't set the magical
"meaning_uri" property on the field. It causes the Python GAE SDK think the
field value is uncompressed.

So if we see a value that is not a JSON dict, assume it is in fact
compressed JSON, and decompress it first.

R=nodir@chromium.org
BUG=801423

Change-Id: I373b960bd7d4b54bc65187f366afbb7b75b72349
Reviewed-on: https://chromium-review.googlesource.com/1072959
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>