[Findit] Classify gtest results in a new format.
To be compatible with json test results, classify gtest results in a new format.

The new method will be used to pre-process test results of test reruns (both deflake and determine pass rate), and then the result will be proccessed differently based on the purpose of the swarming task(This part of changes will be in a following CL).

Bug: 806002
Change-Id: I07e588285b9ffec7d57602eadb154a8c0fb3ba12
Reviewed-on: https://chromium-review.googlesource.com/1067885
Commit-Queue: Chan Li <chanli@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>