[tricium] Simplify template usage
The previous templating logic was good for a website with multiple template
pages, but the plan for the tricium front-end is to use a single main page
with different views depending on path, with routing done on the front-end.

With this in mind, template usage can be simplified, and the templates
directory can be replaced with a single index.html.

This is based on the model of some other apps in the infra repository
such as sheriff-o-matic and tree-status.

Bug: 674478
Change-Id: I582ed1dea702778606bef3f40444feb20d0dca15
Reviewed-on: https://chromium-review.googlesource.com/1074220
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>