[Monorail] Make work_env less UI-centric.
This CL changes work_env to be less UI-centric by renaming variables
to indicate that we expect and instance of MonorailContext instead of
MonorailRequest.  There are still a few methods that only work when a
MonorailRequest is passed in, but those will be changed as we
implement API handlers that call them.

R=jeffcarp@chromium.org

Bug: monorail:2985
Change-Id: I9e6cd635a1b8ebcd9c75932185cfa7b67e036df8
Reviewed-on: https://chromium-review.googlesource.com/1086150
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Jeff Carpenter <jeffcarp@chromium.org>