[Findit] Allow certain DTOs to deserialize from dicts with extra keys.
This change allows a DTO to specify cls._strict_attributes to false to
avoid breaking when the serialization has unrecognized keys. In such
case, the code will not fail, but will log an exception that should be
visible through stackdriver's error reporting and should email the team.

This is desirable when we deserialize information from external systems
which may include new keys without warning.

This would have prevented the outage caused by the introduction of
'isolate_target_name' key in step_metadata. But would have emailed the
team with the details of the failure.

R=chanli,stgao

Change-Id: I66efee5108b9ec1def924320ad55c275fd97a4c4
Reviewed-on: https://chromium-review.googlesource.com/1088146
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>