[Findit] Adjust flake detection data models.
This CL makes the following changes to the data models:
1. Move the step_name and test_name to the flake occurrence model, and
   make step_name, test_name, build_id the key of the flake occurrence
   model.
2. Make luci_project, normalized_step_name and normalized_test_name the
   key of the flake model.

The main reason for this change is that it makes it easier to map a
Flake entity to a monorail issue. In the old design, multiple Flake
entities map to one monorail issue, so when decide whether to
create/update bugs, it requires quering mutiple Flake entities and
aggregating the information from all of their children Flake Occurrence
entities, which is expensive and complicate.

In the new design, it's a one-on-one mapping between Flake entity and
monorail issue.

The reason why I didn't go with this approach at first place is that I
thought it's hard to query the ocurrences of a specific pair of
step_name, test_name. After moving step_name and test_name to the
flake occurrence model, it turns out it's easy to do, and it even
allows querying flake occurrences grouping by step_name and test_name.

Bug: 849462, 845581
Change-Id: I91ce0ea3f45a4f23ab073edbe3b90adf5619216c
Reviewed-on: https://chromium-review.googlesource.com/1089115
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>