[Findit] Get trybot dimensions from swarmbucket api.
This change gets the trybot dimensions dynamically from buildbucket via
the swarmbucket api when the original failure happened on a luci bot.

Note that this does not yet make the tryjob necessarily happen on a
swarming trybot as this is still determined by the static mapping when
triggering the tryjob, but only gets the dimensions (which may be
ignored if the static mapping points to a buildbot trybot)

A later change will make the tryjobs be triggered on swarming if the
matching dimensions are available on Findit's swarming pool and the
static mapping doesn't force buildbot trybot.

R=chanli,stgao,lijeffrey,liaoyuke

Bug: 775326
Change-Id: I9c113b1ff8db28cad6528a887f3ba26bd34a47d7
Reviewed-on: https://chromium-review.googlesource.com/1093297
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>