[Findit] Flake Aanlyzer - Bug template for wrong results
Currently the bug template instructs sheriffs to assign wrong results back
to Findit team. However the bug is still valid for flakiness, even though
the culprit is wrong. The bug should instead be assigned to a developer to
fix the flakiness, and Findit's wrong result should be filed as a separate
bug.

This change clarifies the instructions for sheriffs and creates a bug
template specific to the analysis with the wrong result so a bug against
Findit can be filed in just 2 clicks.

Bug: 846506
Change-Id: I239a420c6e4ddf51e32748693b446d43cd1a967b
Reviewed-on: https://chromium-review.googlesource.com/1089996
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>