[Findit] Flake Analyzer - Set result_status when pipeline completes
Set result status after Flake Analyzer has come to a conclusion successfully.
Note these values should not be set in OnFinalized, as that is called when
pipelines abort which may not be correct (e.g. setting NOT_FOUND_UNTRIAGED
for an analysis that ran into an exception early on would be incorrect for
determining whether the conclusion is correct which is what result_status
represents).

Bug: 850645
Change-Id: If0100301af5c8fa6cc68956ff0b059498fa22838
Reviewed-on: https://chromium-review.googlesource.com/1097795
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>