[Findit] Deprecate the use of `force_buildbot`.
The only remaining use for force_buildbot was to get the mastername
property to use when triggering tryjobs on luci through the
findit_variable builder.

It turns out that for that case we can use the waterfall mastername
instead, doing away with the extra parameter.

R=chanli,stgao

Bug: 853905
Change-Id: I0e536359dfef607cc97154a60b7a78af7eda81d0
Reviewed-on: https://chromium-review.googlesource.com/1105126
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>