[Findit] Implement rules to create or update bugs.
This CL implements the basic rules to create or update bugs:
1. At most 10 issue in any 24h window.
2. An issue can be created or updated at most once every 24h.
3. At least 3 unreported occurrences with different changes within the past 24h.
4. At least 1 occurrence happened within the past 6h.

Note that this CL doesn't implement any code to actually file the bugs,
it just sets the rules. Code to create or update bugs will be in a
separate CL.

There are other advanced features such as PING the bug, re-open a
closed bug and automatically close a bug. They will be implemented in
separate CLs.

Bug: 845581
Change-Id: Iaadce06e03443b66e73264d733b3c8f91f1c9a61
Reviewed-on: https://chromium-review.googlesource.com/1105683
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>